backend/requirements.txt
fastapi
uvicorn
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-genai
python-dotenv
requests
pydantic
motor
certifi